Machine Learning/requirements.txt
Flask==2.2.2
keras==2.10.0
keras==3.3.3
numpy==1.26.4
Pillow==9.4.0
Pillow==10.3.0
